requirements.txt
slack-bolt==1.18.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
python-dotenv==1.0.0
openai==1.3.0
aws-lambda-powertools==2.30.0 